+++ SonghayCore/Xml/XObjectUtility.XAttribute.cs
