+++ SonghayCore/Xml/XObjectUtility.XNode.cs
